refactor(entity): extract boundary and migration helpers
